Separate stateless components and components that use the state
